+++ .gitignore
@@ -185,3 +185,4 @@ fabric.properties
 .idea
 
 *.csv
+config.json
