+++ env_template
@@ -1,0 +1,1 @@
+ETHERSCAN_KEY=
